# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: T4L API
  version: 3.1.0
  description: Canonical API for T4L clients. UUIDs are generated by clients and revisions by the server.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /me/profile:
    get:
      operationId: getProfile
      responses:
        '200':
          description: Current user's synchronized profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    patch:
      operationId: patchProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserProfilePatch' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '409':
          description: Revision conflict; body contains current server profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
  /me/profile/avatar:
    get:
      operationId: getProfileAvatar
      responses:
        '200':
          description: Canonical JPEG profile image
          headers:
            ETag: { schema: { type: string }, description: Quoted avatar revision }
            Cache-Control: { schema: { type: string } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        '404': { description: No profile image }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: putProfileAvatar
      parameters:
        - { name: clientMutationId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: baseRevision, in: query, required: true, schema: { type: integer, format: int64, minimum: 0 } }
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary, maxLength: 2097152 } }
          image/png: { schema: { type: string, format: binary, maxLength: 2097152 } }
          image/webp: { schema: { type: string, format: binary, maxLength: 2097152 } }
      responses:
        '200':
          description: Avatar updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvatarMutationResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '409':
          description: Avatar revision conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvatarMutationResponse' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: deleteProfileAvatar
      parameters:
        - { name: clientMutationId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: baseRevision, in: query, required: true, schema: { type: integer, format: int64, minimum: 0 } }
      responses:
        '200':
          description: Avatar removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvatarMutationResponse' }
        '409':
          description: Avatar revision conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvatarMutationResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /workspaces/{workspaceId}/category-trees:
    get:
      operationId: listCategoryTrees
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Category trees visible in the workspace
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategoryTree' }
  /workspaces/{workspaceId}/reports/aggregate:
    get:
      operationId: aggregateReport
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: true, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Canonical category durations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AggregateRow' }
  /workspaces/{workspaceId}/reports/intersection:
    post:
      operationId: calculateIntersection
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IntersectionRequest' }
      responses:
        '200':
          description: Duration where every selected timeline matches its category filter
          content:
            application/json:
              schema:
                type: object
                required: [durationSeconds]
                properties:
                  durationSeconds: { type: integer, format: int64, minimum: 0 }
  /workspaces/{workspaceId}/reports/distribution:
    get:
      operationId: getActualTimeDistribution
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Actual own and descendant-inclusive duration per category, in milliseconds
          content:
            application/json:
              schema:
                type: object
                required: [from, to, categories]
                properties:
                  from: { type: string, format: date-time }
                  to: { type: string, format: date-time }
                  categories: { type: array, items: { $ref: '#/components/schemas/DetailedDuration' } }
  /workspaces/{workspaceId}/plans/{planId}/distribution:
    get:
      operationId: getPlanTimeDistribution
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - { name: planId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Manual budget and computed timeline distribution of one plan
          content:
            application/json:
              schema:
                type: object
                required: [planId, categories]
                properties:
                  planId: { type: string, format: uuid }
                  categories: { type: array, items: { $ref: '#/components/schemas/PlanDuration' } }
        '404': { description: Plan not found }
  /workspaces/{workspaceId}/tasks/{taskId}/time:
    get:
      operationId: getTaskTime
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Directly tracked task time and signed deviation from estimate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTime' }
        '404': { description: Task not found }
  /workspaces/{workspaceId}/planner/recommendations:
    post:
      operationId: getPlannerRecommendations
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecommendationRequest' }
      responses:
        '200':
          description: Deterministic task recommendations for the available window
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaskRecommendation' }
  /sync/push:
    post:
      operationId: pushMutations
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PushRequest' }
      responses:
        '200':
          description: Per-mutation outcome; retries with the same mutation ID are idempotent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PushResponse' }
  /sync/changes:
    get:
      operationId: pullChanges
      parameters:
        - { name: workspaceId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, required: false, schema: { type: integer, format: int64, minimum: 0, default: 0 } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 200 } }
      responses:
        '200':
          description: Ordered delta page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangePage' }
  /sync/snapshot:
    get:
      operationId: getSnapshot
      parameters:
        - { name: workspaceId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Full replica snapshot for initial or expired-cursor sync
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceSnapshot' }
  /imports:
    post:
      operationId: importWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, snapshot]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                snapshot: { $ref: '#/components/schemas/WorkspaceSnapshot' }
      responses:
        '200':
          description: Imported workspace with remapped IDs
          content:
            application/json:
              schema:
                type: object
                required: [workspaceId, importedEntities]
                properties:
                  workspaceId: { type: string, format: uuid }
                  importedEntities: { type: integer, minimum: 0 }
  /diagnostics/events:
    post:
      operationId: submitClientDiagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiagnosticEventBatch' }
      responses:
        '202': { description: Diagnostic events accepted }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication is required or the bearer token is invalid
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    InvalidRequest:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    PayloadTooLarge:
      description: Request body exceeds the supported limit
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    UserProfile:
      type: object
      required: [revision, avatarRevision, hasAvatar]
      properties:
        birthDate: { type: [string, 'null'], format: date }
        lifeExpectancyYears: { type: [number, 'null'], minimum: 0.1, maximum: 130 }
        revision: { type: integer, format: int64, minimum: 0 }
        avatarRevision: { type: integer, format: int64, minimum: 0 }
        hasAvatar: { type: boolean }
    UserProfilePatch:
      type: object
      required: [clientMutationId, baseRevision, changedFields]
      properties:
        clientMutationId: { type: string, format: uuid }
        baseRevision: { type: integer, format: int64, minimum: 0 }
        changedFields:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [birthDate, lifeExpectancyYears] }
        birthDate: { type: [string, 'null'], format: date }
        lifeExpectancyYears: { type: [number, 'null'], minimum: 0.1, maximum: 130 }
    AvatarMutationResponse:
      type: object
      required: [avatarRevision, hasAvatar]
      properties:
        avatarRevision: { type: integer, format: int64, minimum: 0 }
        hasAvatar: { type: boolean }
...
    CategoryTree:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - type: object
          required: [name, role, sortOrder, archived]
          properties:
            name: { type: string, minLength: 1, maxLength: 120 }
            role: { type: string, enum: [standard, primary, secondary] }
            sortOrder: { type: integer }
            archived: { type: boolean }
            trashedAt: { type: [string, 'null'], format: date-time, readOnly: true }
            purgedAt: { type: [string, 'null'], format: date-time, readOnly: true }
...
    AggregateRow:
      type: object
      required: [categoryTreeId, categoryId, durationSeconds]
      properties:
        categoryTreeId: { type: string, format: uuid }
        categoryId: { type: [string, 'null'], format: uuid }
        durationSeconds: { type: integer, format: int64, minimum: 0 }
    DetailedDuration:
      type: object
      required: [categoryTreeId, categoryId, ownMillis, totalMillis]
      properties:
        categoryTreeId: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        ownMillis: { type: integer, format: int64, minimum: 0 }
        totalMillis: { type: integer, format: int64, minimum: 0 }
    PlanDuration:
      type: object
      required: [categoryTreeId, categoryId, ownMinutes, totalMinutes, timelineOwnMillis, timelineTotalMillis]
      properties:
        categoryTreeId: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        ownMinutes: { type: [integer, 'null'], minimum: 0 }
        totalMinutes: { type: integer, minimum: 0 }
        timelineOwnMillis: { type: integer, format: int64, minimum: 0 }
        timelineTotalMillis: { type: integer, format: int64, minimum: 0 }
    TaskTime:
      type: object
      required: [taskId, spentMillis, estimateMinutes, deviationMillis, calculatedAt]
      properties:
        taskId: { type: string, format: uuid }
        spentMillis: { type: integer, format: int64, minimum: 0 }
        estimateMinutes: { type: integer, minimum: 0 }
        deviationMillis: { type: integer, format: int64 }
        calculatedAt: { type: string, format: date-time }
...
    IntersectionRequest:
      type: object
      required: [from, to, filters]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        filters:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/IntersectionFilter' }
...
    TaskRecommendation:
      type: object
      required: [taskId, suggestedMinutes, score]
      properties:
        taskId: { type: string, format: uuid }
        suggestedMinutes: { type: integer, minimum: 1 }
        score: { type: number }